Feat: new homepage and update redirect routes accordingly and moved main dashboard to its own file
