+++ app/routes/login.tsx
@@ -20,7 +20,7 @@ export const action: ActionFunction = async({ request }) => {
         return json({ error: error.message }, { status: 400 });
     }
 
-    return redirect('/');
+    return redirect('/dashboard');
 }
 
 export default function Login() {
